products_from_libai/telegram-bot-system/requirements.txt
python-telegram-bot==20.7
httpx==0.25.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
